products/pytorch-hyperlight/requirements.txt
pytorch-lightning == 1.1.*, >=1.1.5
tabulate # == 0.8.7
hyperopt # == 0.2.5
tqdm # == 4.54.1
pandas # == 1.1.5
wandb # == 0.10.12
transformers >= 4.1.1
ray[tune] >= 1.1.0
requests # == 2.25.1
torchvision # == 0.8.2
pillow # == 8.1.0 
matplotlib # == 3.3.3
validators # == 0.18.2
boto3 # == 1.16.50
torch # == 1.7.1
torchvision # == 0.8.2
kaggle # == 1.5.10
